requirements.txt
absl-py==2.3.1
aenum==3.1.16
affine==2.4.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
blessings==1.7
branca==0.8.1
cdsapi==0.7.6
certifi==2025.7.14
cftime==1.6.4.post1
charset-normalizer==3.4.2
click==8.2.1
click-plugins==1.1.1.2
cligj==0.7.2
comm==0.2.3
contextily==1.6.2
contourpy==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.15
decorator==5.2.1
ecmwf-datastores-client==0.2.0
ee==0.2
executing==2.2.0
filelock==3.13.1
fiona==1.10.1
flatbuffers==25.2.10
folium==0.20.0
fonttools==4.59.0
fsspec==2024.6.1
gast==0.6.0
geographiclib==2.0
geopandas==0.14.4
geopy==2.4.1
google-pasta==0.2.0
grpcio==1.74.0
h5netcdf==1.6.3
h5py==3.14.0
idna==3.10
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==3.11.0
kiwisolver==1.4.8
libclang==18.1.1
load-dotenv==0.1.0
mapclassify==2.10.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mercantile==1.2.1
ml_dtypes==0.5.3
mpmath==1.3.0
multiurl==0.3.7
mypy_extensions==1.1.0
namex==0.1.0
nest-asyncio==1.6.0
netCDF4==1.7.2
networkx==3.5
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
oauthlib==3.3.1
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
Pygments==2.19.2
pyparsing==3.2.3
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
rasterio==1.4.3
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.1.0
rioxarray==0.19.0
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
sentinelhub==3.10.2
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
snuggs==1.4.7
stack-data==0.6.3
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.6.11
tomli==2.2.1
tomli_w==1.2.0
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
triton==3.3.1
typing-inspect==0.9.0
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.170
ultralytics-thop==2.0.14
urllib3==2.5.0
utm==0.8.1
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xarray==2025.7.1
xyzservices==2025.4.0
